pyproject.toml
[project]
name = "btc-quant"
version = "0.1.0"
description = "Système d'analyse quantitative BTC : ingestion, features, backtest, alertes."
readme = "README.md"
requires-python = ">=3.12,<3.13"
authors = [{ name = "Enzo Paduano" }]
license = { text = "Proprietary" }

# Étape 1 — bootstrap : on ne charge que les libs nécessaires au démarrage.
# Les libs lourdes (data, features, backtest, ml) sont en extras et seront
# activées étape par étape via `uv sync --extra <name>`.
dependencies = [
    "pydantic>=2.7",
    "pydantic-settings>=2.2",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "loguru>=0.7",
]

[project.optional-dependencies]
data = [
    "pybit>=5.7",
    "pandas>=2.2",
    "numpy>=1.26",
    "polars>=1.0",
    "duckdb>=1.0",
    "pyarrow>=16",
]
features = [
    "pandas-ta>=0.3.14b",
    "arch>=7.0",
    "hmmlearn>=0.3.2",
    "ruptures>=1.1.9",
    "pykalman>=0.9.7",
    "scipy>=1.13",
    "scikit-learn>=1.5",
]
backtest = [
    "vectorbt>=0.26",
]
ml = [
    "lightgbm>=4.3",
]

[dependency-groups]
dev = [
    "pytest>=8.2",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "black>=24.0",
    "mypy>=1.10",
    "pre-commit>=3.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["core", "data", "features", "signals", "backtest", "live"]

# ----- Tooling -----------------------------------------------------------------

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E", "F", "W",   # pycodestyle / pyflakes
    "I",             # isort
    "N",             # pep8-naming
    "UP",            # pyupgrade
    "B",             # flake8-bugbear
    "C4",            # comprehensions
    "SIM",           # simplify
    "RET",           # return
    "PL",            # pylint subset
]
ignore = [
    "PLR0913",  # many keyword args in signatures — pandas-style is fine
    "PLR0915",  # many statements — orchestration code legitimately does a lot
    "PLR0912",  # many branches — same rationale
]

[tool.ruff.lint.per-file-ignores]
# Magic values are idiomatic in test assertions.
"tests/**/*.py" = ["PLR2004"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_unreachable = true
ignore_missing_imports = true  # libs sans stubs (pandas_ta, vectorbt, pybit…)

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra --strict-markers --strict-config"
testpaths = ["tests"]
filterwarnings = ["error"]
